chore: remove resource type pages and a lot of refactoring
