--- frontend/src/components/FilterTabs/FilterTabs.constants.js
@@ -1,17 +1,0 @@
-export const tabsData = [
-  {
-    label: "All",
-    filterType: "" /** leave me as an empty string to clear the query */,
-    path: "/resources/all",
-  },
-  {
-    label: "Topics",
-    filterType: "topic",
-    path: "/resources/topics",
-  },
-  {
-    label: "Projects",
-    filterType: "project",
-    path: "/resources/projects",
-  },
-];
