+++ frontend/src/components/ResourceList/ResourceList.js
@@ -1,7 +1,7 @@
 import { useContext, useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import { nanoid } from "nanoid";
-import { Context } from "../../Context";
+import { Context } from "../../appContext";
 import SearchForm from "../SearchForm/SearchForm";
 import { bookmarkIcon, removeBookmarkIcon } from "../../svgs";
 import ResourceSkeleton from "../ResourceSkeleton/ResourceSkeleton";
@@ -16,8 +16,6 @@ function ResourceList() {
     setPageTitle,
     renderedResources,
     setRenderedResources,
-    setResourceType,
-    resourceType,
   } = useContext(Context);
 
   const filterType = useLocation().pathname.split("/")[2];
@@ -33,11 +31,6 @@ function ResourceList() {
 
   useEffect(() => {
     handlePageTitleUpdate();
-    // eslint-disable-next-line
-  }, [filterType]);
-
-  useEffect(() => {
-    setResourceType(filterType);
     // eslint-disable-next-line
   }, [filterType]);
 
@@ -101,7 +94,7 @@ function ResourceList() {
           renderedResources.length === allResources.length || searchTerm.trim()
         }
       >
-        {`Load all ${resourceType === "all" ? "resources" : resourceType} (${
+        {`Load all resources (${
           allResources.length - renderedResources.length
         })`}
       </button>
